Extract the inbetween popover
