+++ packages/base-styles/_z-index.scss
@@ -125,9 +125,6 @@ $z-layers: (
 	// Above the block list and the header.
 	".block-editor-block-popover": 31,
 
-	// Under the block popover (block toolbar).
-	".block-editor-block-list__insertion-point-popover": 28,
-
 	// Show snackbars above everything (similar to popovers)
 	".components-snackbar-list": 100000,
