+++ packages/block-editor/src/components/block-popover/README.md
@@ -1,8 +1,10 @@
-# BlockPopover
+# BlockPopover and BlockPopoverInbetween
 
-The BlockPopover component is a component that allows you to render editor UI by the block (in a popover) but outside the the canvas. This is important to avoid messing up with the style and layout of the block list.
+These two components allows rendering editor UI by the block (in a popover) but outside the the canvas. This is important to avoid messing up with the style and layout of the block list.
 
 For example, it's used to render the contextual block toolbar, the in-between block inserter.
+
+## BlockPopover
 
 ### Props
 
@@ -19,3 +21,21 @@ The client ID of the block representing the bottom position of the popover.
 
 -   Type: `String`
 -   Required: No
+
+## BlockPopoverInbetween
+
+### Props
+
+#### previousClientId
+
+The client ID of the block before the popover.
+
+-   Type: `String`
+-   Required: Yes
+
+#### nextClientId
+
+The client ID of the block after the popover.
+
+-   Type: `String`
+-   Required: Yes
