+++ packages/block-editor/src/components/block-popover/index.js
@@ -8,42 +8,12 @@ import classnames from 'classnames';
  */
 import { Popover } from '@wordpress/components';
 import { getScrollContainer } from '@wordpress/dom';
-import { useRefEffect } from '@wordpress/compose';
 
 /**
  * Internal dependencies
  */
 import { __unstableUseBlockElement as useBlockElement } from '../block-list/use-block-props/use-block-refs';
-
-/**
- * Allow scrolling "through" popovers over the canvas. This is only called for
- * as long as the pointer is over a popover. Do not use React events because it
- * will bubble through portals.
- *
- * @param {Object} scrollableRef
- */
-export function usePopoverScroll( scrollableRef ) {
-	return useRefEffect(
-		( node ) => {
-			if ( ! scrollableRef ) {
-				return;
-			}
-
-			function onWheel( event ) {
-				const { deltaX, deltaY } = event;
-				scrollableRef.current.scrollBy( deltaX, deltaY );
-			}
-			// Tell the browser that we do not call event.preventDefault
-			// See https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/addEventListener#improving_scrolling_performance_with_passive_listeners
-			const options = { passive: true };
-			node.addEventListener( 'wheel', onWheel, options );
-			return () => {
-				node.removeEventListener( 'wheel', onWheel, options );
-			};
-		},
-		[ scrollableRef ]
-	);
-}
+import usePopoverScroll from './use-popover-scroll';
 
 export default function BlockPopover( {
 	clientId,
