+++ packages/block-editor/src/components/block-tools/insertion-point.js
@@ -7,24 +7,16 @@ import classnames from 'classnames';
  * WordPress dependencies
  */
 import { useSelect, useDispatch } from '@wordpress/data';
-import {
-	useCallback,
-	useRef,
-	useMemo,
-	createContext,
-	useContext,
-} from '@wordpress/element';
-import { Popover, __unstableMotion as motion } from '@wordpress/components';
+import { useRef, createContext, useContext } from '@wordpress/element';
+import { __unstableMotion as motion } from '@wordpress/components';
 import { useReducedMotion } from '@wordpress/compose';
-import { isRTL } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
 import Inserter from '../inserter';
 import { store as blockEditorStore } from '../../store';
-import { __unstableUseBlockElement as useBlockElement } from '../block-list/use-block-props/use-block-refs';
-import { usePopoverScroll } from '../block-popover';
+import BlockPopoverInbetween from '../block-popover/inbetween';
 
 export const InsertionPointOpenRef = createContext();
 
@@ -78,108 +70,9 @@ function InsertionPointPopover( {
 			isInserterShown: insertionPoint?.__unstableWithInserter,
 		};
 	}, [] );
-	const previousElement = useBlockElement( previousClientId );
-	const nextElement = useBlockElement( nextClientId );
 	const isVertical = orientation === 'vertical';
-	const style = useMemo( () => {
-		if ( ! previousElement && ! nextElement ) {
-			return {};
-		}
 
-		const previousRect = previousElement
-			? previousElement.getBoundingClientRect()
-			: null;
-		const nextRect = nextElement
-			? nextElement.getBoundingClientRect()
-			: null;
-
-		if ( isVertical ) {
-			return {
-				width: previousElement
-					? previousElement.offsetWidth
-					: nextElement.offsetWidth,
-				height:
-					nextRect && previousRect
-						? nextRect.top - previousRect.bottom
-						: 0,
-			};
-		}
-
-		let width = 0;
-		if ( previousRect && nextRect ) {
-			width = isRTL()
-				? previousRect.left - nextRect.right
-				: nextRect.left - previousRect.right;
-		}
-
-		return {
-			width,
-			height: previousElement
-				? previousElement.offsetHeight
-				: nextElement.offsetHeight,
-		};
-	}, [ previousElement, nextElement ] );
-
-	const getAnchorRect = useCallback( () => {
-		if ( ! previousElement && ! nextElement ) {
-			return {};
-		}
-
-		const { ownerDocument } = previousElement || nextElement;
-
-		const previousRect = previousElement
-			? previousElement.getBoundingClientRect()
-			: null;
-		const nextRect = nextElement
-			? nextElement.getBoundingClientRect()
-			: null;
-
-		if ( isVertical ) {
-			if ( isRTL() ) {
-				return {
-					top: previousRect ? previousRect.bottom : nextRect.top,
-					left: previousRect ? previousRect.right : nextRect.right,
-					right: previousRect ? previousRect.left : nextRect.left,
-					bottom: nextRect ? nextRect.top : previousRect.bottom,
-					ownerDocument,
-				};
-			}
-
-			return {
-				top: previousRect ? previousRect.bottom : nextRect.top,
-				left: previousRect ? previousRect.left : nextRect.left,
-				right: previousRect ? previousRect.right : nextRect.right,
-				bottom: nextRect ? nextRect.top : previousRect.bottom,
-				ownerDocument,
-			};
-		}
-
-		if ( isRTL() ) {
-			return {
-				top: previousRect ? previousRect.top : nextRect.top,
-				left: previousRect ? previousRect.left : nextRect.right,
-				right: nextRect ? nextRect.right : previousRect.left,
-				bottom: previousRect ? previousRect.bottom : nextRect.bottom,
-				ownerDocument,
-			};
-		}
-
-		return {
-			top: previousRect ? previousRect.top : nextRect.top,
-			left: previousRect ? previousRect.right : nextRect.left,
-			right: nextRect ? nextRect.left : previousRect.right,
-			bottom: previousRect ? previousRect.bottom : nextRect.bottom,
-			ownerDocument,
-		};
-	}, [ previousElement, nextElement ] );
-
-	const popoverScrollRef = usePopoverScroll( __unstableContentRef );
 	const disableMotion = useReducedMotion();
-
-	const className = classnames(
-		'block-editor-block-list__insertion-point',
-		'is-' + orientation
-	);
 
 	function onClick( event ) {
 		if ( event.target === ref.current && nextClientId ) {
@@ -202,11 +95,6 @@ function InsertionPointPopover( {
 			hideInsertionPoint();
 		}
 	}
-
-	// Only show the in-between inserter between blocks, so when there's a
-	// previous and a next element.
-	const showInsertionPointInserter =
-		previousElement && nextElement && isInserterShown;
 
 	// Define animation variants for the line element.
 	const horizontalLine = {
@@ -261,7 +149,7 @@ function InsertionPointPopover( {
 			...( ! isVertical ? horizontalLine.rest : verticalLine.rest ),
 			opacity: 1,
 			borderRadius: '2px',
-			transition: { delay: showInsertionPointInserter ? 0.4 : 0 },
+			transition: { delay: isInserterShown ? 0.4 : 0 },
 		},
 		hover: {
 			...( ! isVertical ? horizontalLine.hover : verticalLine.hover ),
@@ -281,27 +169,17 @@ function InsertionPointPopover( {
 		},
 	};
 
-	/* eslint-disable jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events */
-	// While ideally it would be enough to capture the
-	// bubbling focus event from the Inserter, due to the
-	// characteristics of click focusing of `button`s in
-	// Firefox and Safari, it is not reliable.
-	//
-	// See: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#Clicking_and_focus
+	const className = classnames(
+		'block-editor-block-list__insertion-point',
+		'is-' + orientation
+	);
+
 	return (
-		<Popover
-			ref={ popoverScrollRef }
-			noArrow
-			animate={ false }
-			getAnchorRect={ getAnchorRect }
-			focusOnMount={ false }
-			className="block-editor-block-list__insertion-point-popover"
-			// Render in the old slot if needed for backward compatibility,
-			// otherwise render in place (not in the the default popover slot).
-			__unstableSlotName={ __unstablePopoverSlot || null }
-			// Forces a remount of the popover when its position changes
-			// This makes sure the popover doesn't animate from its previous position.
-			key={ nextClientId + '--' + rootClientId }
+		<BlockPopoverInbetween
+			previousClientId={ previousClientId }
+			nextClientId={ nextClientId }
+			__unstablePopoverSlot={ __unstablePopoverSlot }
+			__unstableContentRef={ __unstableContentRef }
 		>
 			<motion.div
 				layout={ ! disableMotion }
@@ -315,16 +193,15 @@ function InsertionPointPopover( {
 				onClick={ onClick }
 				onFocus={ onFocus }
 				className={ classnames( className, {
-					'is-with-inserter': showInsertionPointInserter,
+					'is-with-inserter': isInserterShown,
 				} ) }
 				onHoverEnd={ maybeHideInserterPoint }
-				style={ style }
 			>
 				<motion.div
 					variants={ lineVariants }
 					className="block-editor-block-list__insertion-point-indicator"
 				/>
-				{ showInsertionPointInserter && (
+				{ isInserterShown && (
 					<motion.div
 						variants={ inserterVariants }
 						className={ classnames(
@@ -346,28 +223,18 @@ function InsertionPointPopover( {
 					</motion.div>
 				) }
 			</motion.div>
-		</Popover>
+		</BlockPopoverInbetween>
 	);
-	/* eslint-enable jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events */
 }
 
-export default function InsertionPoint( {
-	children,
-	__unstablePopoverSlot,
-	__unstableContentRef,
-} ) {
+export default function InsertionPoint( { children, ...props } ) {
 	const isVisible = useSelect( ( select ) => {
 		return select( blockEditorStore ).isBlockInsertionPointVisible();
 	}, [] );
 
 	return (
 		<InsertionPointOpenRef.Provider value={ useRef( false ) }>
-			{ isVisible && (
-				<InsertionPointPopover
-					__unstablePopoverSlot={ __unstablePopoverSlot }
-					__unstableContentRef={ __unstableContentRef }
-				/>
-			) }
+			{ isVisible && <InsertionPointPopover { ...props } /> }
 			{ children }
 		</InsertionPointOpenRef.Provider>
 	);
